New bar chart: Easier to change and better look and feel
We drop chartist since it doesn't give us the look and feel we want
(there's no way to diplay each bar value in a bar chart). The new
version is a simple HTML graph where we can fit in both the name
of each module and the size in one bar.

Bug: T133760
Change-Id: Iffc78b7bb0975c16d913cf435f0c47a77271a301