collector/modulecss: Fix "Cannot read property of null" error
Steps to reproduce:
1. Run from console:
  `mw.loader.load('http://example.org/main.css', 'text/css')`
2. Open Performance inspector.

cssFromURL() will crash because it encounters an external
stylesheet which has 'href' but not 'cssRules' (due to same-origin
policy).

On Wikimedia wikis, external stylesheets mostly don't exist (due
to the WMF Privacy policy). But MediaWiki in general does support it.
And even on WMF wikis, Gadgets and user scripts may load external
content after opt-in.

Change-Id: I1f55be4514b99a7f228dec92490e93d8db21d6f2