Use require() instead of global mw.loader.require()
The global/static mw.loader method is mainly for debugging from
the browser console. It is not meant to be used in production code.

Production code should instead use one of:
* the locally bound 'require' exposed to the module execution scope
  if the module is a dependency of this module.
* the local 'require' parameter passed to mw.loader.using(),
  if the module is loaded on-demand.

Right now the production require() does not yet enforce warnings
or errors for loading modules that were not ensured first, but
that is planned for the future.

This use was most likely by mistake because I failed to document
its purpose as being internal.

Change-Id: I15dd097f65c7de01b32f5a8746c0a64b1d1025b0