Verify page title namespace before link is added in the sidebar
Krinkle found the missmatch that we have today when checking if
the performance inspector should be active on a page: When we load
the inspector we verify the page title namespace but we don't do
that when creating the link in the toolbox. This makes the
inspector feel broken (you click on a link and nothing happens).

This fix make sure we only show the link when the title is in
the correct namespace.

Bug: T186680
Bug: T186789
Change-Id: Id55cfc2db16c7c24063c94a3c446e3b467e1de75